+++ tests/v0/stage15/plan/vtable_interning_and_trait_error_tests.rs
@@ -18,6 +18,7 @@
 
 #![cfg(test)]
 
+#![allow(deprecated)]  // Stage 15.15: tests use deprecated format_for_user
 use landin_compiler::compile;
 use landin_compiler::TraitError;